# Restructure map creation, and control inlining for better performance