@Gabriella439 2022-06-23
What do you mean when you say that this change no longer exports `toMap`?  I don't see a change to the export list