@Gabriella439 2022-06-23
I think it's fine to publish a breaking change to remove a utility that was intended for internal use